voiceagent/pyproject.toml
[project]
name = "voiceagent"
version = "0.1.0"
description = "Post-op conversation prioritizer + summarizer"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.0",
  "python-dotenv>=1.0",
  "groq>=0.9.0"
]

[project.scripts]
voiceagent = "voiceagent.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}